Refactored code

+++ vmapcrimes_backend/routes/usersRoutes.js
@@ -2,8 +2,9 @@
 const express = require('express');
 const bcrypt =  require('bcrypt');
 const dotenv= require('dotenv');
-const jwt = require("jsonwebtoken");
 const mongoose = require("mongoose");
+
+// custom imports
 
 //initializing some stuff
 const router = express.Router();
@@ -12,74 +13,15 @@ const Role= require('../models/Roles');
 
 // importing npm provided middlewares
 const { body, validationResult,param } = require('express-validator');
-const { expressjwt: ejwt } = require('express-jwt');
 
 // importing custom middlewares
 const verifyAccess = require('../middleware/verifyAccess');
 
+
 dotenv.config();
 
-// ROUTE 1
-// [ 🟢 UNRESTRICTED ROUTE ]  (May need to later control access with IP based filtering)
-// POST to http://localhost:5001/api/v1/admin/adminLogin with { "email":"test@test.com","password" : "test123" }
 
-router.post('/adminLogin',
-    [
-        body('email').isEmail() //check if the email is valid email or not
-    ],
-
-    async (req,res) => {
-        const errors = validationResult(req);
-        if(!errors.isEmpty()) {
-            return res.status(400).json(errors);
-        }
-        var loginuser
-        try {
-            loginuser = await user.findOne({email: req.body.email}) ;
-        } catch(MongooseError) {
-            return res.status(500).json({status:"failure",message:"Failed to query Database"})
-        }
-        var userRole;
-        try {
-            userRole = await Role.findOne({_id: loginuser.role});
-        }catch(e) {
-            console.log(e)
-            if (e instanceof mongoose.Error) {
-                return res.status(500).json({status:"failure",message:"Failed to query the database"})
-            }else{
-                return res.status(500).json({status:"failure",message:"Some Unknown error occured"})
-
-            }
-        }
-
-        if(userRole.name !== "admin"){
-            return res.status(401).json({status:"failure",message:"User is not an admin!"})
-        }
-        if(!loginuser || !bcrypt.compare(req.body.password,loginuser.password)) {
-            return res.status(404).send("Please enter valid Credentials");
-        }
-    
-
-        // After successful username, password verification
-        var payload =  {
-            id : loginuser._id,
-            username: loginuser.name
-        };
-
-        const jwtSessionTok = await jwt.sign(payload,process.env.JWT_SECRET_KEY,{ algorithm : 'HS256',expiresIn: 2*3600});
-        console.log(jwtSessionTok);
-
-        return res.send(jwtSessionTok);
-    }
-
-
-) 
-// END ROUTE 1
-
-// [ 🔴 NEED ADMIN ACCESS ]
-
-// ROUTE 2
-// POST to http://localhost:5001/api/v1/admin/createUser with { "name" : "John Doe",  "email": "test@test.com","password" : "test123", "phone" : "1234567890" , "address" : "dummy address", "role" : "admin2"}
+// POST to http://localhost:5001/api/user/createUser with { "name" : "John Doe",  "email": "test@test.com","password" : "test123", "phone" : "1234567890" , "address" : "dummy address", "role" : "admin2"}
 
 router.post('/createUser',
 
@@ -165,56 +107,11 @@ router.post('/createUser',
 
 
 // ROUTE 3
-// POST to http://localhost:5001/api/v1/admin/addRole with -> { 'role' : 'normaluser', 'read-perms' : [] , 'action-perms' : [] }
 
-
-router.post('/addRole',
-verifyAccess(),
-    body('role').notEmpty().withMessage("Role name cannot be empty").isLength({min :5, max:20}).withMessage("Role names should be between 5-20 characters in length"),
-async (req,res) => {
-    try {
-    
-        if (req.body.read_perms && !Array.isArray(req.body.read_perms)) {
-            return res.status(400).json({status: 'failure',message:"Read Permissions should be an array of permissions(Strings)"})
-        }
-        if (req.body.action_perms && !Array.isArray(req.body.action_perms)) {
-            return res.status(400).json({status: 'failure',message:"Action Permissions should be an array of permissions(Strings)"})
-        }
-        var roleExists;
-         
-            roleExists = await Role.findOne({name: req.body.role});
-        
-        if(roleExists){
-            return res.status(500).json({status:"failure",message:"Cannot create aRolewith that name, Role exists already"})
-        }
-
-        await Role.create({
-            name: req.body.role,
-            read_perms : req.body.read_perms,
-            action_perms : req.body.action_perms,
-    
-        })
-        
-
-        return res.json({status:"success",message:"New Role Created"})
-    } catch(e) {
-        console.log("Error: "+e)
-
-        if (e instanceof mongoose.Error) {
-            return res.status(500).json({status:"failure",message:"Failed to query Database"})
-
-        }else{
-            
-            return res.status(500).json({status:"failure",message:"Some unknown error occured"})
-
-        }
-    }
-
-}) 
 // END ROUTE 3
 
 // ROUTE 4
-// GET to http://localhost:5001/api/v1/admin/fetchAllUsers 
+// GET to http://localhost:5001/api/user/fetchAllUsers 
 
 router.get('/fetchAllUsers',verifyAccess(), async (req,res) => {
         
@@ -249,7 +146,7 @@ router.get('/fetchAllUsers',verifyAccess(), async (req,res) => {
 
 
 // ROUTE 5
-// GET to http://localhost:5001/api/v1/admin/fetchUser/<userId>
+// GET to http://localhost:5001/api/user/fetchUser/<userId>
 
 router.get('/fetchUser/:id',verifyAccess(),async (req,res) => {
 const id = req.params.id;
@@ -282,13 +179,16 @@ return res.json({_id: _id,name :name,email:email,address:address,phone:phone,rol
 
 
 // ROUTE 6
-// DELETE to http://localhost:5001/api/v1/admin/deleteUser/<userId>
+// DELETE to http://localhost:5001/api/user/deleteUser/<userId>
 
 router.delete('/deleteUser/:id',verifyAccess(),async (req,res) => {  
     if (typeof req.params.id !== 'string') {
         return res.status(400).json({status:"failure",message: "Provide a valid ID"});
     }
     var id = req.params.id.toLowerCase();
+    if(id === req.auth.id) {
+        return res.status(400).json({status:"failure",message:"Unable to delete authenticated user. Please contact an admin"})
+    }
     let toDelete;
     try {
         toDelete = await user.findOneAndDelete({_id: id });
@@ -315,41 +215,11 @@ router.delete('/deleteUser/:id',verifyAccess(),async (req,res) => {
 // END ROUTE 6
 
 // ROUTE 7
-// DELETE to http://localhost:5001/api/v1/admin/deleteRole/<userId>
 
-router.delete('/deleteRole/:id',verifyAccess(),async (req,res) => {  
-    if (typeof req.params.id !== 'string') {
-        return res.status(400).json({status:"failure",message: "Provide a valid ID"});
-    }
-    var id = req.params.id.toLowerCase();
-    let toDelete;
-    try {
-        toDelete = await Role.findOneAndDelete({_id: new mongoose.Types.ObjectId(id) });
-        if(!toDelete){
-            return res.status(404).json({status:"failure",message:"Role not found"})
-        }
-        await user.deleteMany({role: id}) // to delete all the users having the deleted roles
-    } catch(e) {
-        console.log("Error: "+e)
-
-        if (e instanceof mongoose.Error) {
-            return res.status(500).json({status:"failure",message:"Failed to query Database"})
-
-        }else{
-            
-            return res.status(500).json({status:"failure",message:"Some unknown error occured"})
-
-        }
-    }
-        return res.json({status : "success", message : "Role deleted Successfully!"});
-    
-
-
-});
 // END ROUTE 7
 
 // ROUTE 8
-// PUT to http://localhost:5001/api/v1/admin/updateUser
+// PUT to http://localhost:5001/api/user/updateUser
 
 router.put('/updateUser',verifyAccess(),
 [
@@ -374,7 +244,6 @@ router.put('/updateUser',verifyAccess(),
     if (!errors.isEmpty()) {
         return res.status(400).json({errors : errors.array()});
     }
-console.log()
     
     if(typeof req.body._id !== 'string'){
         return res.status(400).json({status:"failure",message:"Please provide a valid id for the user to update"})
@@ -408,148 +277,11 @@ console.log()
 
 
 // ROUTE 9
-// PUT to http://localhost:5001/api/v1/admin/updateRole
 
-router.put('/updateRole',verifyAccess(), 
-[
-    body('name').optional().isLength({min:5,max:20}).withMessage("Name of aRolemust be betweem 5-20 characters long"),
-    
-]
-,
-async (req,res) => {
-    if (req.body.read_perms && !Array.isArray(req.body.read_perms)) {
-        return res.status(400).json({status: 'failure',message:"Read Permissions should be an array of permissions(Strings)"})
-    }
-    if (req.body.action_perms && !Array.isArray(req.body.action_perms)) {
-        return res.status(400).json({status: 'failure',message:"Action Permissions should be an array of permissions(Strings)"})
-    }
-    try {
-        let toUpdate = await Role.findOneAndUpdate({_id: req.body._id},req.body);
-        if(!toUpdate) {
-            return res.status(404).json({status:"failure",message:"Role not found"})
-        }
-    }catch(e) {
-        console.log("Error: "+e)
-
-        if (e instanceof mongoose.Error) {
-            return res.status(500).json({status:"failure",message:"Failed to query Database"})
-
-        }else{
-            
-            return res.status(500).json({status:"failure",message:"Some unknown error occured"})
-
-        }
-    }
-
-    return res.json({status:"success",message:"User updated Successfully"});
-})
-// END ROUTE 9
-
-// ROUTE 10
-// GET to http://localhost:5001/api/v1/admin/fetchRole/<roleID>
-
-router.get('/fetchRole/:id',verifyAccess(), 
-async (req,res) => {
-        // console.log(req.params.id+"length is "+req.params.id.length)
-        var lengthChecks= (req.params.id.length==24)
-        // console.log("length checks "+lengthChecks)
-        var isAlphanumeric =  /^[a-zA-Z0-9]+$/.test(req.auth.id)
-        // console.log("is alpha numeric "+isAlphanumeric)
-
-        if (!req.params.id || !(typeof req.params.id === 'string')|| !lengthChecks || !isAlphanumeric) {
-            return res.status(400).json({status:"failure",message : "Please Send a valid id in parameter"})
-         }
-
-         try {
-            let toFetch = await Role.find({_id: req.params.id.toLowerCase()});
-            if (!toFetch) {
-                return res.status(404).json({status:"failure",message:"Role not found"})
-            }
-            return res.json(toFetch[0])
-
-         } catch(e) {
-            console.log("Error: "+e)
-    
-            if (e instanceof mongoose.Error) {
-                return res.status(500).json({status:"failure",message:"Failed to query Database"})
-    
-            }else{
-                
-                return res.status(500).json({status:"failure",message:"Some unknown error occured"})
-    
-            }
-        }
-
-    }
-)
-// END ROUTE 10
-
-// ROUTE 11
-// GET to http://localhost:5001/api/v1/admin/fetchAllRoles
-
-router.get('/fetchAllRoles',verifyAccess(), async (req,res) => {
-    try {
-        let fetchArray = await Role.find();
-        return res.json({roles:fetchArray})
-    }catch(e) {
-        console.log("Error: "+e)
-
-        if (e instanceof mongoose.Error) {
-            return res.status(500).json({status:"failure",message:"Failed to query Database"})
-
-        }else{
-            
-            return res.status(500).json({status:"failure",message:"Some unknown error occured"})
-
-        }
-    }
-})
-// END ROUTE 11
-
-// ROUTE 12
-// DELETE to http://localhost:5001/api/v1/admin/deleteRoles with an array of ids in the body
-
-router.delete('/deleteRoles',
-verifyAccess(),async (req,res) => {  
-    
-    if(!(Array.isArray(req.body.ids))) {
-        return res.status(400).json({status: "failure",message: "Please send ids as array"});
-    }
-    if(!(req.body.ids.every((item) => {
-        return (typeof item === "string" && item.length == 24) 
-    }))) {
-        return res.status(400).json({status: "failure",message: "Bad Request"})
-    }
-
-    let toDelete;
-    try {
-        toDelete = awaitRoledeleteMany({_id: {$in:req.body.ids} });
-        if(!(toDelete.deletedCount === req.body.ids.length)){
-            await user.deleteMany({role: {$in:req.body.ids}}) // to delete all the users having the deleted roles
-            return res.status(500).json({status:"failure",message:"Something went wrong, All roles could not be deleted"})
-        }
-        await user.deleteMany({role: {$in:req.body.ids}}) // to delete all the users having the deleted roles
-    } catch(e) {
-        console.log("Error: "+e)
-
-        if (e instanceof mongoose.Error) {
-            return res.status(500).json({status:"failure",message:"Failed to query Database"})
-
-        }else{
-            
-            return res.status(500).json({status:"failure",message:"Some unknown error occured"})
-
-        }
-    }
-        return res.json({status : "success", message : "Role deleted Successfully!"});
-    
-
-
-});
 // END ROUTE 12
 
 // ROUTE 13
-// DELETE to http://localhost:5001/api/v1/admin/deleteUsers with a body containing an array of ids to delete
+// DELETE to http://localhost:5001/api/user/deleteUsers with a body containing an array of ids to delete
 
 router.delete('/deleteUser/:id',verifyAccess(),async (req,res) => {  
     if(!(Array.isArray(req.body.ids))) {
@@ -584,5 +316,28 @@ router.delete('/deleteUser/:id',verifyAccess(),async (req,res) => {
 
 
 });
-// END ROUTE 13
-module.exports = router;
+
+// GET to /api/user/fetchUserCount
+
+router.get("/fetchUserCount",verifyAccess(),async (req,res) => 
+{
+    console.log("In fetch user count")
+    try{
+        count = await user.countDocuments({})
+   } catch(e) {
+       console.log("Error: "+e)
+
+       if (e instanceof mongoose.Error) {
+           return res.status(500).json({status:"failure",message:"Failed to query Database"})
+
+       }else{
+           
+           return res.status(500).json({status:"failure",message:"Some unknown error occured"})
+
+       }
+   }
+   return res.json({count})
+   }
+)
+
+module.exports = router
